dom/tests/mochitest/webcomponents/test_custom_element_callback_innerhtml.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1102502
-->
<head>
  <title>Test for connected callback for element created in the document by the parser</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1102502">Bug 1102502</a>
<div id="container"></div>

<script>

SimpleTest.waitForExplicitFinish();

var connectedCallbackCount = 0;

class Foo extends HTMLElement {
  connectedCallback() {
    ok(true, "connectedCallback should be called when the parser creates an element in the document.");
    connectedCallbackCount++;
    // connectedCallback should be called twice, once for the element created for innerHTML and
    // once for the element created in this document.
    if (connectedCallbackCount == 2) {
      SimpleTest.finish();
    }
  }
};

customElements.define("x-foo", Foo);

var container = document.getElementById("container");
container.innerHTML = '<x-foo></x-foo>';

</script>

<x-foo></x-foo>

</body>
</html>
